Adjust all tests
